refactor ros2 and cover with tests
